casa-dia - 12/12

+++ src/controllers/livrosController.js
@@ -1,3 +1,4 @@
+import NaoEncontrado from "../erros/NaoEncontrado.js";
 import livros from "../models/Livro.js";
 
 class LivroController {
@@ -17,12 +18,16 @@ class LivroController {
   static listarLivroPorId = async (req, res, next) => {
     try {
       const id = req.params.id;
-
-      const livroResultados = await livros.findById(id)
+      const livroResultado = await livros.findById(id)
         .populate("autor", "nome")
         .exec();
 
-      res.status(200).send(livroResultados);
+      if (livroResultado !== null) {
+        res.status(200).send(livroResultado);
+      } else {
+        next(new NaoEncontrado("Id do Livro não encontrado."));
+      }
+
     } catch (error) {
       next(error);
     }
@@ -44,9 +49,17 @@ class LivroController {
     try {
       const id = req.params.id;
 
-      await livros.findByIdAndUpdate(id, {$set: req.body});
+      const livroResultado = await livros.findById(id)
+        .populate("autor", "nome")
+        .exec();
 
-      res.status(200).send({message: "Livro atualizado com sucesso"});
+        if (livroResultado !== null) {
+          await livros.findByIdAndUpdate(id, { $set: req.body });
+          res.status(200).send({ message: "Livro atualizado com sucesso" });
+        } else {
+          next(new NaoEncontrado("Id do Livro não encontrado."));
+        }
+
     } catch (error) {
       next(error);
     }
@@ -55,10 +68,16 @@ class LivroController {
   static excluirLivro = async (req, res, next) => {
     try {
       const id = req.params.id;
+      const livroResultado = await livros.findById(id)
+        .populate("autor", "nome")
+        .exec();
 
-      await livros.findByIdAndDelete(id);
-
-      res.status(200).send({message: "Livro removido com sucesso"});
+      if (livroResultado !== null) {
+        await livros.findByIdAndDelete(id);
+        res.status(200).send({ message: "Livro removido com sucesso" });
+      } else {
+        next(new NaoEncontrado("Id do Livro não encontrado."));
+      }
     } catch (error) {
       next(error);
     }
@@ -68,9 +87,13 @@ class LivroController {
     try {
       const editora = req.query.editora;
 
-      const livrosResultado = await livros.find({"editora": editora});
-
-      res.status(200).send(livrosResultado);
+      const livrosResultado = await livros.find({ "editora": editora });
+      
+       if (livrosResultado !== null) {
+        res.status(200).send(livrosResultado);
+      } else {
+        next(new NaoEncontrado("Editora não encontrada."));
+      }
     } catch (error) {
       next(error);
     }
